updating isPfelement logic
If the --type parameter is set during the yo pfelement call, the type question is never asked and the default value of the type question is not used. This fix will use the type from options if it is provided.

Also, we need to make sure we include the dot files for the elements created in the PatternFly Elements repository.